face_recognition/requirements.txt
numpy
pillow
requests
flask
torchvision
# torch
